# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD API",
        "version": "1.0.0",
        "description": "API documentation for CRUD operations on items and categories"
    },
    "tags": [
        {
            "name": "Items",
            "description": "Endpoints related to item management"
        },
        {
            "name": "Categories",
            "description": "Endpoints related to category management"
        }
    ],
    "paths": {
        "/api/items": {
            "get": {
                "tags": ["Items"],
                "summary": "Get all items",
                "responses": {
                    "200": { "description": "Successful response" },
                    "500": { "description": "Server error" }
                }
            },
            "post": {
                "tags": ["Items"],
                "summary": "Create a new item",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Item" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Item created successfully" },
                    "400": { "description": "Validation error" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": ["Items"],
                "summary": "Get a single item by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": { "description": "Successful response" },
                    "404": { "description": "Item not found" },
                    "500": { "description": "Server error" }
                }
            },
            "put": {
                "tags": ["Items"],
                "summary": "Update an item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Item" }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Item updated successfully" },
                    "400": { "description": "Validation error" },
                    "404": { "description": "Item not found" },
                    "500": { "description": "Server error" }
                }
            },
            "delete": {
                "tags": ["Items"],
                "summary": "Delete an item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": { "description": "Item deleted successfully" },
                    "404": { "description": "Item not found" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/items/bulk": {
            "post": {
                "tags": ["Items"],
                "summary": "Create multiple items",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": { "$ref": "#/components/schemas/Item" }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Items created successfully" },
                    "400": { "description": "Validation error" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": ["Categories"],
                "summary": "Get all categories",
                "responses": {
                    "200": { "description": "Successful response" },
                    "500": { "description": "Server error" }
                }
            },
            "post": {
                "tags": ["Categories"],
                "summary": "Create a new category",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Category" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Category created successfully" },
                    "400": { "description": "Validation error" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": ["Categories"],
                "summary": "Get a single category by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": { "description": "Successful response" },
                    "404": { "description": "Category not found" },
                    "500": { "description": "Server error" }
                }
            },
            "put": {
                "tags": ["Categories"],
                "summary": "Update a category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Category" }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Category updated successfully" },
                    "400": { "description": "Validation error" },
                    "404": { "description": "Category not found" },
                    "500": { "description": "Server error" }
                }
            },
            "delete": {
                "tags": ["Categories"],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": { "description": "Category deleted successfully" },
                    "404": { "description": "Category not found" },
                    "500": { "description": "Server error" }
                }
            }
        },
        "/api/categories/bulk": {
            "post": {
                "tags": ["Categories"],
                "summary": "Create multiple categories",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": { "$ref": "#/components/schemas/Category" }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Categories created successfully" },
                    "400": { "description": "Validation error" },
                    "500": { "description": "Server error" }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Item": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "price": { "type": "number" }
                },
                "required": ["name", "description", "price"]
            },
            "Category": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "description": { "type": "string" }
                },
                "required": ["name", "description"]
            }
        }
    }
}